+++ lib/noticed.rb
@@ -4,6 +4,7 @@ require "noticed/engine"
 module Noticed
   autoload :Base, "noticed/base"
   autoload :Coder, "noticed/coder"
+  autoload :Model, "noticed/model"
   autoload :Translation, "noticed/translation"
 
   module DeliveryMethods