+++ test/dummy/db/schema.rb
@@ -1,6 +1,21 @@
-ActiveRecord::Schema.define(version: 2020_07_26_201932) do
+# This file is auto-generated from the current state of the database. Instead
+# of editing this file, please use the migrations feature of Active Record to
+# incrementally modify your database, and then regenerate this schema definition.
+#
+# This file is the source Rails uses to define your schema when running `rails
+# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
+#
+# It's strongly recommended that you check this file into your version control system.
+
+ActiveRecord::Schema.define(version: 2020_08_03_191250) do
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
   create_table "notifications", force: :cascade do |t|
-    t.bigint "account_id"
+    t.integer "account_id"
     t.string "recipient_type", null: false
     t.bigint "recipient_id", null: false
     t.string "type"
@@ -11,7 +26,6 @@ ActiveRecord::Schema.define(version: 2020_07_26_201932) do
     t.index ["account_id"], name: "index_notifications_on_account_id"
     t.index ["recipient_type", "recipient_id"], name: "index_notifications_on_recipient_type_and_recipient_id"
   end
-
   create_table "accounts", force: :cascade do |t|
     t.string "name"
     t.datetime "created_at", precision: 6, null: false